#33. 인가(권한 확인)

+++ app/Http/Controllers/ArticleController.php
@@ -12,7 +12,15 @@ class ArticleController extends Controller
         return view('articles/create');
     }
 
-    public function store(Request $request) {
+    public function store(Request $request, Article $article) {
+        // User 모델을 활용하여 어플리케이션의 권한 확인 => 권한이 없을 때 동작을 지정.
+        if (!Auth::user()->can('update', $article)) {
+            abort(403);
+        }
+
+        // 컨트롤러 헬퍼를 활용하여 어플리케이션의 권한 확인 => 권한이 없으면 자동으로 응답 만들어줌.
+        // $this->authorize('update', $article);
+
         $input = $request->validate([
             'body' => [
                 'required',
@@ -44,6 +52,9 @@ class ArticleController extends Controller
     }
 
     public function edit(Article $article) {
+        if (!Auth::user()->can('update', $article)) {
+            abort(403);
+        }
         return view('articles.edit', ['articles' => $article]);
     }
 
@@ -63,6 +74,10 @@ class ArticleController extends Controller
     }
 
     public function destroy(Article $article) {
+        if (!Auth::user()->can('update', $article)) {
+            abort(403);
+        }
+
         $article->delete();
 
         return redirect()->route('articles.index');
